+++ pnpm-workspace.yaml.genie.ts
@@ -3,5 +3,6 @@ import { rootWorkspacePackages } from './package.json.genie.ts'
 
 export default pnpmWorkspaceYaml.root({
   packages: rootWorkspacePackages,
+  repoName: 'effect-utils',
   ...commonPnpmWorkspaceData,
 })